+++ docs-src/src/pages/premium.tsx
@@ -484,16 +484,6 @@ export default function Premium() {
                                                                 Fulltext Search
                                                             </a>
                                                         </li>
-                                                        <li>
-                                                            <a href="/reactivity.html" target="_blank">
-                                                                Reactivity Vue
-                                                            </a>
-                                                        </li>
-                                                        <li>
-                                                            <a href="/reactivity.html" target="_blank">
-                                                                Reactivity Preact Signals
-                                                            </a>
-                                                        </li>
                                                     </ul>
                                                 </div>
                                             </div>
@@ -999,26 +989,6 @@ export default function Premium() {
                                             <h4>FlexSearch</h4>
                                             <p>
                                                 A plugin to efficiently run local fulltext search indexing and queries.
-                                            </p>
-                                        </div>
-                                    </div>
-                                </a>
-                                <a href="/reactivity.html" target="_blank">
-                                    <div className="premium-block hover-shadow-middle bg-gradient-left-top">
-                                        <div className="premium-block-inner">
-                                            <h4>Reactivity Vue</h4>
-                                            <p>
-                                                An extension for Vue.js to get vue shallow-ref objects to observe RxDB state instead of rxjs observables.
-                                            </p>
-                                        </div>
-                                    </div>
-                                </a>
-                                <a href="/reactivity.html" target="_blank">
-                                    <div className="premium-block hover-shadow-middle bg-gradient-right-top">
-                                        <div className="premium-block-inner">
-                                            <h4>Reactivity Preact Signals</h4>
-                                            <p>
-                                                An extension for react/preact to get preact signals to observe RxDB state instead of rxjs observables.
                                             </p>
                                         </div>
                                     </div>